noveldeglace:  - refactored for plugin beta  - releaseDate as Date.toString()  - separted parseNovels  - separted getCheerio  - loadedCheerio => $ for shorter lines & simpler test on web
